Update active_storage_overview.md
Trying to add a weirdly hard to find fact.

The [ActiveStorage::Attachment](https://edgeapi.rubyonrails.org/classes/ActiveStorage/Attachment.html) states this fact in the first paragraph, which is very odd that it is the _only_ place I was able to find, which states that.

In fact, I think the description of the `def purge` method in `ActiveStorage::Blob` is bad, as it actually doesn't state this fact but instead claims that calling it will delete the file. The Blobs `before_destroy` is actually preventing it to be deleted, in case there is any attachment still referencing it.